[AIRFLOW-5901] Group tests for the tasks and variables command (#6554)
